refactor: abstract facets to its own component
removes code duplication
