+++ src/components/DoiContainer/DoiContainer.tsx
@@ -16,9 +16,7 @@ import {
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faEnvelope } from '@fortawesome/free-solid-svg-icons'
 import { faTwitter, faFacebook } from '@fortawesome/free-brands-svg-icons'
-import { faSquare, faCheckSquare } from '@fortawesome/free-regular-svg-icons'
-import { useRouter } from 'next/router'
-import Link from 'next/link'
+import DoiFacet from '../DoiFacet/DoiFacet'
 
 type Props = {
   item?: string
@@ -201,7 +199,6 @@ interface DoiQueryVar {
 
 const DoiContainer: React.FunctionComponent<Props> = ({ item }) => {
   // const [selectedOption, setSelectedOption] = React.useState('')
-  const router = useRouter()
 
   const [doi, setDoi] = React.useState<DoiType>()
   const [cursor] = useQueryState('cursor', { history: 'push' })
@@ -262,125 +259,6 @@ const DoiContainer: React.FunctionComponent<Props> = ({ item }) => {
 
   if (error) {
     return <Error title="No Content" message="Unable to retrieve Content" />
-  }
-
-  const renderFacets = () => {
-    if (loading) return <div className="col-md-3"></div>
-
-    if (!loading && doi.citations.nodes.length == 0)
-      return <div className="col-md-3"></div>
-
-    function facetLink(param: string, value: string) {
-      let url = '?'
-      let icon = faSquare
-
-      // get current query parameters from next router
-      let params = new URLSearchParams(router.query as any)
-
-      // delete person parameter
-      params.delete('doi')
-
-      // delete cursor parameter
-      params.delete('cursor')
-
-      if (params.get(param) == value) {
-        // if param is present, delete from query and use checked icon
-        params.delete(param)
-        icon = faCheckSquare
-      } else {
-        // otherwise replace param with new value and use unchecked icon
-        params.set(param, value)
-      }
-
-      url += params.toString()
-      return (
-        <Link href={url}>
-          <a>
-            <FontAwesomeIcon icon={icon} />{' '}
-          </a>
-        </Link>
-      )
-    }
-
-    return (
-      <div className="panel panel-transparent">
-        <div className="panel facets add">
-          <div className="panel-body">
-            <h4>Publication Year</h4>
-            <ul id="published-facets">
-              {doi.citations.published.map((facet) => (
-                <li key={facet.id}>
-                  {facetLink('published', facet.id)}
-                  <div className="facet-title">{facet.title}</div>
-                  <span className="number pull-right">
-                    {facet.count.toLocaleString('en-US')}
-                  </span>
-                  <div className="clearfix" />
-                </li>
-              ))}
-            </ul>
-          </div>
-        </div>
-
-        <div className="panel facets add">
-          <div className="panel-body">
-            <h4>Work Type</h4>
-            <ul id="work-type-facets">
-              {doi.citations.resourceTypes.map((facet) => (
-                <li key={facet.id}>
-                  {facetLink('resource-type', facet.id)}
-                  <div className="facet-title">{facet.title}</div>
-                  <span className="number pull-right">
-                    {facet.count.toLocaleString('en-US')}
-                  </span>
-                  <div className="clearfix" />
-                </li>
-              ))}
-            </ul>
-          </div>
-        </div>
-
-        {doi.citations.registrationAgencies && doi.citations.registrationAgencies.length > 0 && (
-          <div className="panel facets add">
-            <div className="panel-body">
-              <h4>Registration Agency</h4>
-              <ul id="repository-facets">
-                {doi.citations.registrationAgencies.map((facet) => (
-                  <li key={facet.id}>
-                    {facetLink('registration-agency', facet.id)}
-                    <div className="facet-title">{facet.title}</div>
-                    <span className="number pull-right">
-                      {facet.count.toLocaleString('en-US')}
-                    </span>
-                    <div className="clearfix" />
-                  </li>
-                ))}
-              </ul>
-            </div>
-          </div>
-        )}
-
-        {doi.citations.languages && doi.citations.languages.length > 0 && (
-          <div className="panel facets add">
-            <div className="panel-body">
-              <h4>Language</h4>
-              <ul id="affiliation-facets">
-                {doi.citations.languages.map((facet) => (
-                  <li key={facet.id}>
-                    {facetLink('language', facet.id)}
-                    <div className="facet-title">{facet.title}</div>
-                    <span className="number pull-right">
-                      {facet.count.toLocaleString('en-US')}
-                    </span>
-                    <div className="clearfix" />
-                  </li>
-                ))}
-              </ul>
-            </div>
-          </div>
-        )}
-      </div>
-    )
   }
 
   const leftSideBar = () => {
@@ -510,7 +388,7 @@ const DoiContainer: React.FunctionComponent<Props> = ({ item }) => {
             </span>
           </div>
         </div>
-        {renderFacets()}
+        <DoiFacet model="doi" data={doi.citations} loading={loading}></DoiFacet>
       </div>
     )
   }
